feat: split GenomeAuthoritySystem based on modules
The integration with Genome and BasicCrafting is optional for SimpleFarming.

As a component system can only be loaded properly if all it's dependencies are also active, this means that the combined `GenomeAuthoritySystem` is only loaded if **both** Genome and BasicCrafting are active.

By splitting the authority system into `GenomeExtensionAuthoritySystem` and `GenomeCraftingExtensionAuthoritySystem` each extension system can run with the minimal set of required modules.

- `GenomeExtensionAuthoritySystem` (works if at least Genome is loaded)
- `GenomeCraftingExtensionAuthoritySystem` (works if both Genome and Basic Crafting are loaded)

Fixes #119
